refactor(hooks): 훅 책임 분리 및 파일 단위 구조화
- useAuthApi를 세션·좋아요·프로필·계정 4개 훅 파일로 분리
- useGameDetailModal을 useGameDetailData + useGameDetailLikeAction으로 분리
- useMainPageGames에서 쿼리 로직과 유틸을 별도 파일로 추출
